refactor: remove remaining "nexus" branding from codebase
Rename app/components/nexus/ -> app/components/portfolio/, hero
section id="nexus" -> id="home" (and matching nav anchors/default
active state), Cloudinary asset prefix nexus_profile_ -> profile_.
Historical v3.0.0 changelog entry in README kept as-is since it
accurately describes that era's actual "Nexus" branding.
