+++ app/components/portfolio/FaviconUpdater.tsx
