+++ app/components/portfolio/Footer.tsx
